@@ -6,8 +6,17 @@
     "type": "module",
     "main": "./dist/index.js",
     "module": "./dist/index.js",
-    "exports": "./dist/index.js",
-    "types": "./dist/index.d.ts",
+    "types": "./dist/types.d.ts",
+    "exports": {
+        ".": {
+            "types": "./dist/types.d.ts",
+            "import": "./dist/index.js"
+        },
+        "./github-actions": {
+            "types": "./dist/exports-for-compare-action.d.ts",
+            "import": "./dist/exports-for-compare-action.js"
+        }
+    },
     "files": [
         "dist"
     ],
@@ -22,9 +31,8 @@
     "scripts": {
         "prebuild": "rm -rf ./dist",
         "build": "tsc --project tsconfig.prod.json",
-        "start": "node dist --config eslint-remote-tester.config.cjs",
+        "start": "node dist --config eslint-remote-tester.config.js",
         "start:memory-limit-crash": "NODE_OPTIONS=--max_old_space_size=50 node dist",
-        "lint": "eslint . --max-warnings 0 --ext .js,.ts,.tsx && publint",
         "test:integration": "vitest run --config test/integration/vitest.config.integration.ts",
         "test:smoke": "vitest run --config test/smoke/vitest.config.smoke.ts",
         "validate": "pnpm build && pnpm lint && pnpm test && pnpm test:integration"
@@ -45,29 +53,34 @@
         "simple-git": "^3.24.0"
     },
     "devDependencies": {
+        "@eslint/eslintrc": "^3.1.0",
         "@types/babel__code-frame": "^7.0.6",
         "@types/eslint": "^8.56.10",
         "@types/node": "^20.12.12",
         "@types/object-hash": "^3.0.6",
         "@types/react": "^17.0.80",
-        "@typescript-eslint/eslint-plugin": "^7.11.0",
-        "@typescript-eslint/parser": "^7.11.0",
         "conventional-changelog-cli": "^5.0.0",
-        "eslint": "^8.57.0",
+        "eslint": "^9.3.0",
+        "eslint-plugin-local-rules": "^2.0.1",
         "eslint-remote-tester-repositories": "workspace:*",
+        "importx": "^0.3.5",
         "ink-testing-library": "^2.1.0",
+        "jiti": "^1.21.3",
         "node-pty": "^1.0.0",
         "strip-ansi": "^6.0.1",
-        "ts-node": "^10.9.2",
         "typescript": "^5.4.5",
         "vitest": "^1.6.0"
     },
     "peerDependencies": {
-        "eslint": ">=7",
-        "ts-node": ">=9.0.0"
+        "eslint": ">=9",
+        "importx": ">=0.3.5",
+        "jiti": ">=1"
     },
     "peerDependenciesMeta": {
-        "ts-node": {
+        "importx": {
+            "optional": true
+        },
+        "jiti": {
             "optional": true
         }
     },
